# Pre-compiled regexp optimization