# Pre-compiled regexp optimization

Improve performance by precompiling regular expressions that are used repeatedly.

### COSBench

|Op-Type	|Op-Count	|Byte-Count	|Avg-ResTime	|Avg-ProcTime	|Throughput	|Bandwidth	|Succ-Ratio | 
|---|---|---|---|---|---|---|---|
|BASELINE op1: read	| 294.63 kops	| 2.95 GB	| 232.11 ms	| 232.11 ms	| 327.42 op/s	| 3.27 MB/S	| 100%
| PR op1: read	|339.79 kops	|3.4 GB	|201.28 ms	|201.28 ms	|377.58 op/s	|3.78 MB/S	|100%


https://bugzilla.redhat.com/show_bug.cgi?id=2106742

Signed-off-by: Alexander Indenbaum <aindenba@redhat.com>