@guymguym 2022-11-24
@baum can you also write what is the op size in the benchmarks you made above? and what is the backing store/namespace you used. thanks