@baum 2022-11-24
@guymguym the test is performed on RGW backing store, see https://bugzilla.redhat.com/show_bug.cgi?id=2106742

COSBench workstage definition

```yaml
        <workstage name="randread" closuredelay="0" config="">
            <auth type="none"/>
            <storage type="s3" config="accesskey=XXXXX;secretkey=YYYYY;endpoint=http://s3-openshift-storage.apps.bm.example.com;path_style_access=true"/>
            <work name="randread" type="normal" workers="76"
                interval="5" division="object" runtime="900" rampup="0"
                rampdown="0" afr="200000" totalOps="0" totalBytes="0" config="">
                <auth type="none"/>
                <operation type="read" ratio="100" division="object"
                    config="cprefix=tccmx7n2nq6gmrgz5x92;containers=u(1,100);objects=u(1,1000);hashCheck=true" id="op1"/>
            </work>
        </workstage>
```