@guymguym 2022-12-01
@baum that read stage doesn't say the size - I think this is a result of the previous stages where it create objects by some size distribution. But if I divide the byte-count by the ops-count and get an average of 10 KB - sounds about right?